Prepare file store backend split
